Refactor: codebase folder struct cleanup
