+++ src/state/rollup_account_loader.rs
@@ -1,0 +1,42 @@
+use std::sync::RwLock;
+use std::collections::HashMap;
+use solana_client::rpc_client::RpcClient;
+use solana_sdk::{
+    pubkey::Pubkey,
+    account::AccountSharedData,
+};
+use solana_svm::transaction_processing_callback::TransactionProcessingCallback;
+use solana_sdk::account::ReadableAccount;
+
+pub struct RollUpAccountLoader<'a> {
+    cache: RwLock<HashMap<Pubkey, AccountSharedData>>,
+    rpc_client: &'a RpcClient,
+}
+
+impl<'a> RollUpAccountLoader<'a> {
+    pub fn new(rpc_client: &'a RpcClient) -> Self {
+        Self {
+            cache: RwLock::new(HashMap::new()),
+            rpc_client,
+        }
+    }
+}
+
+// impl InvokeContextCallback for RollUpAccountLoader<'_> {}
+impl TransactionProcessingCallback for RollUpAccountLoader<'_> {
+    fn get_account_shared_data(&self, pubkey: &Pubkey) -> Option<AccountSharedData> {
+        if let Some(account) = self.cache.read().unwrap().get(pubkey) {
+            return Some(account.clone());
+        }
+
+        let account: AccountSharedData = self.rpc_client.get_account(pubkey).ok()?.into();
+        self.cache.write().unwrap().insert(*pubkey, account.clone());
+
+        Some(account)
+    }
+
+    fn account_matches_owners(&self, account: &Pubkey, owners: &[Pubkey]) -> Option<usize> {
+        self.get_account_shared_data(account)
+            .and_then(|account| owners.iter().position(|key| account.owner().eq(key)))
+    }
+}
